+++ src/main/java/searchengine/repositories/SiteRepository.java
@@ -13,6 +13,8 @@ public interface SiteRepository extends JpaRepository<SiteEntity, Long> {
 
 	SiteEntity findByUrl(String url);
 
+	void deleteById(Integer id);
+
 }