+++ src/main/java/searchengine/services/RepositoryService.java
@@ -42,6 +42,10 @@ public interface RepositoryService {
 
 	void deleteIndexes();
 
+	void deleteSiteById(Integer Id);
+
+	void deletePageById(Integer id);
+
 	void deleteAllTables();
 
 	void flushRepositories();